rename: common 폴더 @common으로 네이밍 변경
