+++ src/components/@common/SideNavbar/index.tsx
@@ -1,6 +1,6 @@
 import { styled, Stack } from "@mui/material";
 import LogoIcon from "@/assets/logo.svg?react";
-import AccordionLinkListButton from "@/components/common/SideNavbar/AccordionLinkListButton";
+import AccordionLinkListButton from "@/components/@common/SideNavbar/AccordionLinkListButton";
 import {
   historyLinkButtonInfoList,
   membersLinkButtonInfoList,
