+++ src/pages/PaymentStatusPage.tsx
@@ -1,6 +1,6 @@
 import PaymentStatusHeader from "@/components/PaymentStatus/PaymentStatusHeader";
 import PaymentStatusInfoTable from "@/components/PaymentStatus/PaymentStatusInfoTable";
-import Title from "@/components/common/Title";
+import Title from "@/components/@common/Title";
 import PaymentStatusContextProvider from "@/contexts/PaymentStatusContext";
 
 export default function PaymentStatusPage() {
