+++ src/pages/PendingMembersPage.tsx
@@ -1,6 +1,6 @@
 import PendingMembersHeader from "@/components/PendingMembers/PendingMembersHeader";
 import PendingMembersInfoTable from "@/components/PendingMembers/PendingMembersInfoTable";
-import Title from "@/components/common/Title";
+import Title from "@/components/@common/Title";
 import PendingMembersContextProvider from "@/contexts/PendingMembersContext";
 
 export default function PendingMembersPage() {
