+++ src/pages/RecruitmentPage.tsx
@@ -1,6 +1,6 @@
 import RecruitmentHeader from "@/components/Recruitment/RecruitmentHeader";
 import RecruitmentInfoTable from "@/components/Recruitment/RecruitmentInfoTable";
-import Title from "@/components/common/Title";
+import Title from "@/components/@common/Title";
 import RecruitmentContextProvider from "@/contexts/RecruitmentContext";
 
 export default function RecruitmentPage() {
